+++ src/components/AirdropWalletsCarousel.tsx
@@ -4,7 +4,7 @@ import Carousel from 'react-native-snap-carousel';
 
 import { AirdropWalletDetails } from 'app/consts';
 
-import { AirdropWalletsCarouselItem } from './AirdropWalletsCarouselItem';
+import { AirdropWalletBalanceCard } from './AirdropWalletBalanceCard';
 
 interface Props {
   items: AirdropWalletDetails[];
@@ -14,19 +14,12 @@ interface Props {
 const SCREEN_WIDTH = Dimensions.get('window').width;
 
 export class AirdropWalletsCarousel extends Component<Props> {
-  carouselRef = React.createRef<any>();
-
-  snap = (index: number) => {
-    this.carouselRef.current!.snapToItem(index, true);
-  };
-
   render() {
     return (
       <Carousel
         containerCustomStyle={this.props.styles}
         data={this.props.items}
-        ref={this.carouselRef}
-        renderItem={({ item }: { item: AirdropWalletDetails }) => <AirdropWalletsCarouselItem walletDetails={item} />}
+        renderItem={({ item }: { item: AirdropWalletDetails }) => <AirdropWalletBalanceCard walletDetails={item} />}
         sliderWidth={SCREEN_WIDTH}
         itemWidth={SCREEN_WIDTH * 0.56}
       />
